requirements.txt
albumentations==1.4.1
einops
imageio
matplotlib
natsort
numpy
omegaconf
opencv-python==4.10.0.84
Pillow
pyiqa==0.1.11
pytorch-msssim
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX
torch==1.12.0
torchvision==0.13.0
tqdm
